refactor: remove configuration file support
